Move test to buildtagged test file
